Lift out base item fixture for ide_completion tests
